katana_single_app/requirements.txt
pyTelegramBotAPI
openai
python-dotenv
pydantic
pytest-asyncio
